api/requirements.txt
fastapi #Framework python pour l'API
uvicorn[standard] #Serveur pour FastAPI
psycopg2-binary #Driver PostgreSQL pour python